# Comprehensive Sync Loss Optimization with Advanced Quantized Lip Sync Models