… docs/ADVANCED_LIP_SYNC_MODELS.md
… docs/SYNC_LOSS_OPTIMIZATION.md
… src/lip_sync/advanced_smart_selector.py
… src/lip_sync/emo_model.py
… src/lip_sync/enhanced_smart_selector.py
… src/lip_sync/gaussian_splatting_model.py
… src/lip_sync/latentsync_model.py
… src/lip_sync/musetalk_model.py
… src/lip_sync/quantized_models.py
… src/lip_sync/sync_evaluator.py
… src/lip_sync/vasa1_model.py
… test_sync_loss_benchmark.py